+++ brostar_api/urls.py
@@ -48,6 +48,10 @@ urlpatterns = [
     path("redoc/", schema_view.with_ui("redoc", cache_timeout=0), name="schema-redoc"),
     path("api/", include(("api.urls", "api"), namespace="api")),
     path("prometheus/", include("django_prometheus.urls")),
+    path(
+        "duplicaten/",
+        include("shapefile_duplicates.urls", namespace="shapefile_duplicates"),
+    ),
 ]